requirements.txt
boto3==1.19.12
python-dotenv==0.19.1
